+++ src/lib/firestoreFunctions.js
@@ -1,13 +1,5 @@
+// eslint-disable-next-line import/no-unresolved
 import { collection, addDoc } from 'https://www.gstatic.com/firebasejs/9.6.11/firebase-firestore.js';
 import { db } from './config.js';
 
-try {
-    const docRef = await addDoc(collection(db, "users"), {
-      first: "Ada",
-      last: "Lovelace",
-      born: 1815
-    });
-    console.log("Document written with ID: ", docRef.id);
-  } catch (e) {
-    console.error("Error adding document: ", e);
-  }
+export const saveTask = (description) => addDoc(collection(db, 'tasks'), { description });
